+++ server/fly-away/src/main/java/com/codestates/flyaway/web/login/dto/LoginDto.java
@@ -1,0 +1,11 @@
+package com.codestates.flyaway.web.login.dto;
+
+import lombok.Getter;
+
+@Getter
+public class LoginDto {
+
+
+
+
+}
